Module genco::lang

Language specialization for genco

This module contains sub-modules which provide implementations of the Lang trait to configure genco for various programming languages.

This module also provides a dummy Lang implementation for ().

This allows () to be used as a quick and dirty way to do formatting, usually for examples.

use genco::prelude::*;

let tokens: Tokens = quote!(hello world);

Re-exports

pub use self::c::C;

pub use self::csharp::Csharp;

pub use self::dart::Dart;

pub use self::go::Go;

pub use self::java::Java;

pub use self::js::JavaScript;

pub use self::python::Python;

pub use self::rust::Rust;

pub use self::swift::Swift;

Modules

c

Specialization for C code generation.

csharp

Specialization for Csharp code generation.

dart

Specialization for Dart code generation.

go

Specialization for Go code generation.

java

Specialization for Java code generation.

js

Specialization for JavaScript code generation.

python

Specialization for Python code generation.

rust

Specialization for Rust code generation.

swift

Specialization for Swift code generation.

Traits

Lang

Trait to implement for language specialization.

LangItem

A type-erased holder for language-specific items.

LangSupportsEval

Marker trait indicating that a language supports quoted string interpolation.

Functions

c_family_write_quoted

Escape the given string according to a C-family escape sequence.