Module genco::prelude

Prelude containing typical things to import when using the library.

Re-exports

pub use crate::lang::*;

Macros

quote

Whitespace sensitive quasi-quoting.

quote_fn

Convenience macro for constructing a FormatInto implementation in-place.

quote_in

Behaves the same as quote! while quoting into an existing token stream with <target> => <quoted>.

Structs

Tokens

A stream of tokens.

Traits

FormatInto

Trait for types that can be formatted in-place into a token stream.

Functions

display

Function to build a string literal.

quoted

Function to provide string quoting.

register

Function to provide item registration.