Enum genco::tokens::Item

pub enum Item<L>where
    L: Lang,{
    Literal(ItemStr),
    Lang(usize, Box<L::Item>),
    Register(usize, Box<L::Item>),
    Push,
    Line,
    Space,
    Indentation(i16),
    OpenQuote(bool),
    CloseQuote,
    OpenEval,
    CloseEval,
}

A single item in a stream of tokens.

Variants

Literal(ItemStr)

A literal item. Is added as a raw string to the stream of tokens.

Lang(usize, Box<L::Item>)

A language-specific item.

Register(usize, Box<L::Item>)

A language-specific item that is not rendered.

Push

Push a new line unless the current line is empty. Will be flushed on indentation changes.

Line

Push a line. Will be flushed on indentation changes.

Space

Space between language items. Typically a single space.

Multiple spacings in sequence are collapsed into one. A spacing does nothing if at the beginning of a line.

Indentation(i16)

Manage indentation.

An indentation of 0 has no effect.

OpenQuote(bool)

Switch to handling input as a quote.

The argument indicates whether the string contains any interpolated values.

The string content is quoted with the language-specific [quoting method]. [quoting method]: Lang::Openquote_string

CloseQuote

Close the current quote.

OpenEval

Switch on evaluation. Only valid during string handling.

CloseEval

Close evaluation.