Trait genco::tokens::FormatInto

pub trait FormatInto<L>
where
    L: Lang,
{
    // Required method
    fn format_into(self, tokens: &mut Tokens<L>);
}

Trait for types that can be formatted in-place into a token stream.

Things implementing FormatInto can be used as arguments for interpolation in the quote! macro.

from_fn() is a helper function which simplifies the task of creating a FormatInto implementation on the fly.

Examples

use genco::quote_in;
use genco::tokens::{ItemStr, FormatInto, from_fn, static_literal};
use genco::lang::Lang;

fn comment<L>(s: impl Into<ItemStr>) -> impl FormatInto<L>
where
    L: Lang
{
    from_fn(move |tokens| {
        let s = s.into();
        quote_in!(*tokens => $(static_literal("//")) $s);
    })
}

Required Methods

fn format_into(self, tokens: &mut Tokens<L>)

Convert the type into tokens in-place.

A simple way to build ad-hoc format_into implementations is by using the from_fn() function.